IN2006-2250 ensure pdf with layers is shown
If a pdf with layers is created and uploaded to moodle, the layers are not shown. But the layers are there in the original. One can convert the pdf as images, then the layers are shown.

There is a tracker entry for this as well: https://tracker.moodle.org/browse/MDL-64431